fix card image changing slow due to large image size
